+++ src/pages/news/component/NewsArticle/Top.tsx
@@ -1,7 +1,7 @@
 export const NewsArticleTop = (props: any) => {
   return (
     <div>
-      <h1 className="md:p-8 text-xl md:text-2xl font-bold text-center">{props.mainTitle}</h1>
+      <h1 className="md:p-8 text-4xl md:text-4xl font-bold text-left ">{props.mainTitle}</h1>
       <p className="py-5 pr-5 text-right text-gray-600">{props.updatedAt}</p>
     </div>
   );
